Move gaussian test.
